move folder view to src
