+++ src/View/Home/Home.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Slider from '../../Slider/Slider';
+import Slider from '../../components/Slider/Slider';
 import './home.css';
 
 function Home() {
